Updating the SDK to remove all global variables and wrap the global namespace functions into the `Mapcode` namespace. Also updating all the unit test stuff, samples, docs, etc. to match the new stuff.
Additionally, added a `composer.json` file so the package can be loaded via the package manager now.
